Refactor/create custom hook
