+++ src/api/post.js
@@ -1,10 +1,5 @@
 import supabaseClient from 'lib/supabaseClient';
 
-const getPosts = async () => {
-  const { data } = await supabaseClient.from('posts').select('*');
-  return data;
-};
-
 const getDetail = async (id) => {
   const { data } = await supabaseClient.from('posts').select().eq('id', id);
   return data;
@@ -18,14 +13,8 @@ const updatePost = async (post) => {
   await supabaseClient.from('posts').update(post).eq('id', post.id);
 };
 
-const getLikes = async (id) => {
-  // const { data } = await supabaseClient.from('likes').select().eq('postId', id);
-  // const { data } = await supabaseClient.from('posts').select('likes').eq('id', id);
-  // return data;
-};
-
 const updateLikes = async (post) => {
   await supabaseClient.from('posts').update(post).eq('id', post.id);
 };
 
-export { getPosts, getDetail, deletePost, updatePost, getLikes, updateLikes };
+export { getDetail, deletePost, updatePost, updateLikes };
